- Refactoring of validation rules to be more precise
