+++ blobporter.go
@@ -1,22 +1,17 @@
 package main
 
 import (
-	"errors"
 	"flag"
 	"fmt"
 	"log"
 	"math"
-	"net/url"
 	"os"
 	"runtime"
 	"strconv"
-	"strings"
 	"sync/atomic"
 	"time"
 
 	"github.com/Azure/blobporter/pipeline"
-	"github.com/Azure/blobporter/sources"
-	"github.com/Azure/blobporter/targets"
 	"github.com/Azure/blobporter/transfer"
 	"github.com/Azure/blobporter/util"
 )
@@ -170,16 +165,20 @@ func displayFinalWrapUpSummary(duration time.Duration, targetRetries int32, thre
 
 func main() {
 
-	parseAndValidate()
+	flag.Parse()
 
 	// Create pipelines
-	sourcePipelines, targetPipeline := getPipelines()
+	sourcePipelines, targetPipeline, err := getPipelines()
+
+	if err != nil {
+		log.Fatal(err)
+	}
+
 	stats := transfer.NewStats(numberOfWorkers, numberOfReaders)
 
 	for b, sourcePipeline := range sourcePipelines {
 		sourcesInfo := sourcePipeline.GetSourcesInfo()
 
-		adjustReaders(len(sourcesInfo))
 		tfer := transfer.NewTransfer(&sourcePipeline, &targetPipeline, numberOfReaders, numberOfWorkers, blockSize)
 
 		displayFilesToTransfer(sourcesInfo, len(sourcePipelines), b)
@@ -194,432 +193,6 @@ func main() {
 
 	stats.DisplaySummary()
 
-}
-
-const openFileLimitForLinux = 1024
-
-//validates if the number of readers needs to be adjusted to accommodate max filehandle limits in Debian systems
-func adjustReaders(numOfSources int) {
-	//var source []pipeline.SourcePipeline
-	//var target pipeline.TargetPipeline
-	if runtime.GOOS != "linux" {
-		return
-	}
-
-	if transferType == transfer.FileToBlock || transferType == transfer.FileToPage {
-		if (numOfSources * numberOfHandlesPerFile) > openFileLimitForLinux {
-			numberOfHandlesPerFile = openFileLimitForLinux / (numOfSources * numberOfHandlesPerFile)
-
-			if numberOfHandlesPerFile == 0 {
-				log.Fatal("The number of files will cause the process to exceed the limit of open files allowed by the OS. Reduce the number of files to be transferred")
-			}
-
-			fmt.Printf("Warning! Adjusting the number of handles per file (-h) to %v\n", numberOfHandlesPerFile)
-
-		}
-	}
-}
-
-func isSourceHTTP() bool {
-
-	url, err := url.Parse(sourceURIs[0])
-
-	return err == nil && (strings.ToLower(url.Scheme) == "http" || strings.ToLower(url.Scheme) == "https")
-}
-
-func getTransferPipelines(pipelinesFactory transferPipelinesFactory, validationRules ...parseAndvalidationRule) ([]pipeline.SourcePipeline, pipeline.TargetPipeline) {
-	for i := 0; i < len(validationRules); i++ {
-		val := validationRules[i]
-		if err := val(); err != nil {
-			log.Fatal(err)
-		}
-	}
-
-	return pipelinesFactory()
-}
-
-type transferPipelinesFactory func() ([]pipeline.SourcePipeline, pipeline.TargetPipeline)
-
-func getPipelines() ([]pipeline.SourcePipeline, pipeline.TargetPipeline) {
-
-	switch transferType {
-	case transfer.FileToPage:
-		return getTransferPipelines(getFileToPagePipelines,
-			pvBlobAuthInfoIsReq,
-			pvSourceURIISReq,
-			pvBlockSizeCheckForPageBlobs,
-			pvContainerIsReq)
-	case transfer.FileToBlock:
-		return getTransferPipelines(getFileToBlockPipelines,
-			pvBlobAuthInfoIsReq,
-			pvSourceURIISReq,
-			pvBlockSizeCheckForBlockBlobs,
-			pvContainerIsReq)
-	case transfer.HTTPToPage:
-		return getTransferPipelines(getHTTPToPagePipelines,
-			pvBlobAuthInfoIsReq,
-			pvSourceURIISReq,
-			pvIsSourceHTTP,
-			pvBlockSizeCheckForPageBlobs,
-			pvContainerIsReq)
-	case transfer.HTTPToBlock:
-		return getTransferPipelines(getHTTPToBlockPipelines,
-			pvBlobAuthInfoIsReq,
-			pvSourceURIISReq,
-			pvIsSourceHTTP,
-			pvContainerIsReq,
-			pvBlockSizeCheckForBlockBlobs)
-	case transfer.BlobToFile:
-		return getTransferPipelines(getBlobToFilePipelines,
-			pvBlobAuthInfoIsReq,
-			pvContainerIsReq,
-			pvBlockSizeCheckForBlockBlobs)
-	case transfer.HTTPToFile:
-		return getTransferPipelines(getHTTPToFilePipelines,
-			pvSourceURIISReq,
-			pvIsSourceHTTP)
-	case transfer.BlobToBlock:
-		return getTransferPipelines(getBlobToBlockPipelines,
-			pvBlobAuthInfoIsReq,
-			pvContainerIsReq,
-			pvSourceURIISReq,
-			pvSourceInfoForBlobIsReq,
-			pvBlockSizeCheckForBlockBlobs)
-	case transfer.BlobToPage:
-		return getTransferPipelines(getBlobToPagePipelines,
-			pvBlobAuthInfoIsReq,
-			pvContainerIsReq,
-			pvSourceURIISReq,
-			pvSourceInfoForBlobIsReq,
-			pvBlockSizeCheckForPageBlobs)
-	}
-
-	log.Fatal(fmt.Errorf("Invalid transfer type: %v ", transferType))
-
-	return nil, nil
-}
-func getFileToPagePipelines() (source []pipeline.SourcePipeline, target pipeline.TargetPipeline) {
-
-	params := &sources.MultiFileParams{
-		SourcePatterns:   sourceURIs,
-		BlockSize:        blockSize,
-		TargetAliases:    blobNames,
-		NumOfPartitions:  numberOfReaders,
-		MD5:              calculateMD5,
-		KeepDirStructure: keepDirStructure,
-		FilesPerPipeline: numberOfFilesInBatch}
-
-	source = sources.NewMultiFile(params)
-	target = targets.NewAzurePage(storageAccountName, storageAccountKey, containerName)
-	return
-}
-func getFileToBlockPipelines() (source []pipeline.SourcePipeline, target pipeline.TargetPipeline) {
-	//Since this is default value detect if the source is http and change th transfer type
-	if isSourceHTTP() {
-		source = []pipeline.SourcePipeline{sources.NewHTTP(sourceURIs, blobNames, calculateMD5)}
-		transferDefStr = transfer.HTTPToBlock
-	} else {
-		params := &sources.MultiFileParams{
-			SourcePatterns:   sourceURIs,
-			BlockSize:        blockSize,
-			TargetAliases:    blobNames,
-			NumOfPartitions:  numberOfReaders,
-			MD5:              calculateMD5,
-			KeepDirStructure: keepDirStructure,
-			FilesPerPipeline: numberOfFilesInBatch}
-
-		source = sources.NewMultiFile(params)
-	}
-	target = targets.NewAzureBlock(storageAccountName, storageAccountKey, containerName)
-	return
-}
-func getHTTPToPagePipelines() (source []pipeline.SourcePipeline, target pipeline.TargetPipeline) {
-	source = []pipeline.SourcePipeline{sources.NewHTTP(sourceURIs, blobNames, calculateMD5)}
-	target = targets.NewAzurePage(storageAccountName, storageAccountKey, containerName)
-	return
-}
-func getHTTPToBlockPipelines() (source []pipeline.SourcePipeline, target pipeline.TargetPipeline) {
-	source = []pipeline.SourcePipeline{sources.NewHTTP(sourceURIs, blobNames, calculateMD5)}
-	target = targets.NewAzureBlock(storageAccountName, storageAccountKey, containerName)
-	return
-}
-func getBlobToPagePipelines() (source []pipeline.SourcePipeline, target pipeline.TargetPipeline) {
-
-	params := &sources.AzureBlobParams{
-		Container:         sourceKeyInfo["CONTAINER"],
-		BlobNames:         blobNames,
-		AccountName:       sourceKeyInfo["ACCOUNT_NAME"],
-		AccountKey:        sourceAuthorization,
-		CalculateMD5:      calculateMD5,
-		UseExactNameMatch: exactNameMatch,
-		FilesPerPipeline:  numberOfFilesInBatch,
-		//default to always true so blob names are kept
-		KeepDirStructure: true}
-
-	source = sources.NewAzureBlob(params)
-	target = targets.NewAzurePage(storageAccountName, storageAccountKey, containerName)
-	return
-}
-func getBlobToBlockPipelines() (source []pipeline.SourcePipeline, target pipeline.TargetPipeline) {
-
-	blobNames = []string{sourceKeyInfo["PREFIX"]}
-
-	params := &sources.AzureBlobParams{
-		Container:         sourceKeyInfo["CONTAINER"],
-		BlobNames:         blobNames,
-		AccountName:       sourceKeyInfo["ACCOUNT_NAME"],
-		AccountKey:        sourceAuthorization,
-		CalculateMD5:      calculateMD5,
-		UseExactNameMatch: exactNameMatch,
-		FilesPerPipeline:  numberOfFilesInBatch,
-		//default to always true so blob names are kept
-		KeepDirStructure: true}
-
-	source = sources.NewAzureBlob(params)
-	target = targets.NewAzureBlock(storageAccountName, storageAccountKey, containerName)
-	return
-}
-func getBlobToFilePipelines() (source []pipeline.SourcePipeline, target pipeline.TargetPipeline) {
-
-	if len(blobNames) == 0 {
-		//use empty prefix to get the bloblist
-		blobNames = []string{""}
-	}
-
-	params := &sources.AzureBlobParams{
-		Container:         containerName,
-		BlobNames:         blobNames,
-		AccountName:       storageAccountName,
-		AccountKey:        storageAccountKey,
-		CalculateMD5:      calculateMD5,
-		UseExactNameMatch: exactNameMatch,
-		FilesPerPipeline:  numberOfFilesInBatch,
-		KeepDirStructure:  keepDirStructure}
-
-	source = sources.NewAzureBlob(params)
-	target = targets.NewMultiFile(true, numberOfHandlesPerFile)
-	return
-}
-func getHTTPToFilePipelines() (source []pipeline.SourcePipeline, target pipeline.TargetPipeline) {
-	source = []pipeline.SourcePipeline{sources.NewHTTP(sourceURIs, blobNames, calculateMD5)}
-
-	var targetAliases []string
-	var err error
-
-	if len(blobNames) > 0 {
-		targetAliases = blobNames
-	} else {
-		targetAliases = make([]string, len(sourceURIs))
-		for i, src := range sourceURIs {
-			targetAliases[i], err = util.GetFileNameFromURL(src)
-			if err != nil {
-				log.Fatal(err)
-			}
-		}
-	}
-
-	target = targets.NewMultiFile(true, numberOfHandlesPerFile)
-	return
-}
-
-// parseAndValidate - extra checks on command line arguments
-func parseAndValidate() {
-
-	flag.Parse()
-	var err error
-	if transferType, err = transfer.ParseTransferDefinition(transferDefStr); err != nil {
-		log.Fatal(err)
-	}
-
-	if numberOfFilesInBatch < 1 {
-		log.Fatal("Invalid value for option -x, the value must be greater than 1")
-	}
-
-	if numberOfHandlesPerFile < 1 {
-		log.Fatal("Invalid value for option -h, the value must be greater than 1")
-	}
-
-	if util.HTTPClientTimeout < 30 {
-		fmt.Printf("Warning! The storage HTTP client timeout is too low (>5). Setting value to 600s \n")
-		util.HTTPClientTimeout = 600
-	}
-
-	
-	dedupeLevel, err = transfer.ParseDupeCheckLevel(dedupeLevelOptStr)
-	if err != nil {
-		log.Fatalf("Duplicate detection level is invalid.  Found '%s', must be one of %s", dedupeLevelOptStr, transfer.DupeCheckLevelStr)
-	}
-
-}
-
-//Validation rules...
-type parseAndvalidationRule func() error
-
-func pvBlockSizeCheckForBlockBlobs() error {
-	var err error
-	blockSize, err = util.ByteCountFromSizeString(blockSizeStr)
-	if err != nil {
-		return fmt.Errorf("Invalid block size specified: %v. Parse Error: %v ", blockSizeStr, err)
-	}
-
-	blockSizeMax := util.LargeBlockSizeMax
-	if blockSize > blockSizeMax {
-		return fmt.Errorf("Block size specified (%v) exceeds maximum of %v", blockSizeStr, util.PrintSize(blockSizeMax))
-	}
-
-	return nil
-
-}
-
-func pvIsSourceHTTP() error {
-	if !isSourceHTTP() {
-		return fmt.Errorf("The source is an invalid HTTP endpoint")
-	}
-	return nil
-}
-
-func pvBlockSizeCheckForPageBlobs() error {
-	if blockSize%uint64(targets.PageSize) != 0 {
-		return fmt.Errorf("Invalid block size (%v) for a page blob. The size must be a multiple of %v (bytes) and less or equal to %v (4MB)", blockSize, targets.PageSize, 4*util.MB)
-	}
-
-	if blockSize > 4*util.MB {
-		//adjust the block size to 4 MB
-		blockSize = 4 * util.MB
-
-	}
-
-	return nil
-}
-
-func pvSourceURIISReq() error {
-	if len(sourceURIs) == 0 || sourceURIs == nil {
-		return fmt.Errorf("The source parameter is missing (-f).\nMust be a file, URL, Azure Blob URL or file pattern (e.g. /data/*.fastq) ")
-	}
-	return nil
-}
-
-func pvSourceInfoForBlobIsReq() error {
-	burl, err := url.Parse(sourceURIs[0])
-
-	if err != nil {
-		return fmt.Errorf("Invalid Blob URL. Parsing error: %v", err)
-	}
-
-	host := strings.Split(burl.Hostname(), ".")
-
-	sourceAccountName := host[0]
-
-	if sourceAccountName == "" {
-		return fmt.Errorf("Invalid source Azure Blob URL. Account name could be parsed from the domain")
-	}
-
-	segments := strings.Split(burl.Path, "/")
-
-	sourceContName := segments[1]
-
-	if sourceContName == "" {
-		return fmt.Errorf("Invalid source Azure Blob URL. A container is required")
-	}
-	sourceBlobName := ""
-	if len(segments) > 1 {
-		sourceBlobName = strings.Join(segments[2:len(segments)], "/")
-	}
-
-	sourceKeyInfo = make(map[string]string)
-
-	sourceKeyInfo["ACCOUNT_NAME"] = sourceAccountName
-	sourceKeyInfo["CONTAINER"] = sourceContName
-	sourceKeyInfo["PREFIX"] = sourceBlobName
-
-	if sourceAuthorization == "" {
-		envVal := os.Getenv(sourceAuthorizationEnvVar)
-		if envVal == "" {
-			return fmt.Errorf("The source storage acccount key is required for this transfer type")
-		}
-		sourceAuthorization = envVal
-	}
-
-	return nil
-}
-
-func pvContainerIsReq() error {
-	if containerName == "" {
-		return fmt.Errorf("container name not specified ")
-	}
-	return nil
-}
-
-func pvBlobAuthInfoIsReq() error {
-
-	// wasn't specified, try the environment variable
-	if storageAccountName == "" {
-		envVal := os.Getenv(storageAccountNameEnvVar)
-		if envVal == "" {
-			return errors.New("storage account name not specified or found in environment variable " + storageAccountNameEnvVar)
-		}
-		storageAccountName = envVal
-	}
-
-	// wasn't specified, try the environment variable
-	if storageAccountKey == "" {
-		envVal := os.Getenv(storageAccountKeyEnvVar)
-		if envVal == "" {
-			return errors.New("storage account key not specified or found in environment variable " + storageAccountKeyEnvVar)
-		}
-		storageAccountKey = envVal
-	}
-
-	return nil
-}
-
-//validates command line params considering the type of transfer
-func validateTransferTypesParams() error {
-	var transt transfer.Definition
-	var err error
-	if transt, err = transfer.ParseTransferDefinition(transferDefStr); err != nil {
-		return err
-	}
-
-	if transt == transfer.HTTPToPage || transt == transfer.FileToPage {
-		if blockSize%uint64(targets.PageSize) != 0 {
-			return fmt.Errorf("Invalid block size (%v) for a page blob. The size must be a multiple of %v (bytes) and less or equal to %v (4MB)", blockSize, targets.PageSize, 4*util.MB)
-		}
-
-		if blockSize > 4*util.MB {
-			//adjust the block size to 4 MB
-			blockSize = 4 * util.MB
-
-		}
-	}
-
-	isUpload := transt != transfer.BlobToFile && transt != transfer.HTTPToFile
-
-	if (len(sourceURIs) == 0 || sourceURIs == nil) && isUpload {
-		log.Fatal("The file parameter is missing. Must be a file, URL or file pattern (e.g. /data/*.fastq) ")
-	}
-
-	if transt != transfer.HTTPToFile {
-		// wasn't specified, try the environment variable
-		if storageAccountName == "" {
-			envVal := os.Getenv(storageAccountNameEnvVar)
-			if envVal == "" {
-				return errors.New("storage account name not specified or found in environment variable " + storageAccountNameEnvVar)
-			}
-			storageAccountName = envVal
-		}
-
-		// wasn't specified, try the environment variable
-		if storageAccountKey == "" {
-			envVal := os.Getenv(storageAccountKeyEnvVar)
-			if envVal == "" {
-				return errors.New("storage account key not specified or found in environment variable " + storageAccountKeyEnvVar)
-			}
-			storageAccountKey = envVal
-		}
-	}
-
-	return nil
 }
 
 func getProgressBarDelegate(totalSize uint64, quietMode bool) func(r pipeline.WorkerResult, committedCount int, bufferLevel int) {
